Moved ServerEulaService to proper place
